+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.9.22
 
+- Add new early game weapon: Wizard Staff
 - Fix luminance of spell projectiles and effects
 
 # 0.9.21